safari-600.7-branch/LayoutTests/canvas/philip/tests/toDataURL.lowercase.unicode.html

<!DOCTYPE html>
<title>Canvas test: toDataURL.lowercase.unicode</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

// Use LATIN CAPITAL LETTER I WITH DOT ABOVE (Unicode lowercase is "i")
var data = canvas.toDataURL('\u0130mage/png');
_assertMatch(data, /^data:image\/png[;,]/, "data", "/^data:image\\/png[;,]/");

var data = canvas.toDataURL('\u0130mage/jpeg');
_assertMatch(data, /^data:image\/png[;,]/, "data", "/^data:image\\/png[;,]/");


});
</script>
