LayoutTests/canvas/philip/tests/toDataURL.lowercase.unicode-expected.txt

< [index] >
toDataURL.lowercase.unicode
toDataURL type is ASCII-case-insensitive
References: toDataURL.lowercase
Actual output:
Passed
